requirements.txt
flask==3.0.2
pandas==2.2.1
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.1
gunicorn==21.2.0 